wip - data model changes
